Bloqueio de auto exclusão e limpeza de código

+++ resources/views/components/table.blade.php
@@ -25,44 +25,15 @@
                                     @endforeach
                                     <td>
                                         <a href="javascript: void(0);" class="text-reset fs-16 px-1"> <i class="ri-settings-3-line"></i></a>
-                                        {{-- <div style="display: inline-block" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="danger-tooltip" data-bs-title="Excluir registro">
-                                            <a href="javascript: console.log({{ $item->id }});" class="text-reset fs-16 px-1" data-bs-toggle="modal" data-bs-target="#modal-exclusao"> 
-                                                <i class="ri-delete-bin-2-line"></i>
-                                            </a>
-                                        </div> --}}
-                                        <div style="display: inline-block" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="danger-tooltip" data-bs-title="Excluir registro">
-                                            <a href="javascript: deleteModal({{ $item->id }});" class="text-reset fs-16 px-1"> 
-                                                <i class="ri-delete-bin-2-line"></i>
-                                            </a>
-                                        </div>
-                                        {{-- <div style="display: inline-block" data-bs-toggle="modal" data-bs-target="#modal-exclusao{{$item->id}}" >
-                                            <a href="javascript: void(0)"  class="text-reset fs-16 px-1" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="danger-tooltip" data-bs-title="Excluir registro"> 
-                                                <i class="ri-delete-bin-2-line"></i>
-                                            </a>
-                                            <div id="modal-exclusao{{$item->id}}" class="modal fade" data-bs-backdrop="static" tabindex="-1" role="dialog" aria-labelledby="fill-danger-modalLabel" aria-hidden="true">
-                                                <div class="modal-dialog">
-                                                    <div class="modal-content modal-filled bg-danger">
-                                                        <div class="modal-header">
-                                                            <h4 class="modal-title" id="fill-danger-modalLabel">CUIDADO!</h4>
-                                                            <button type="button" class="btn-close btn-close-white" data-bs-dismiss="modal" aria-label="Close"></button>
-                                                        </div>
-                                                        <div class="modal-body">
-                                                            <p>Você realmente tem certeza que deseja excluir esse registro?</p>
-                                                            <p>A informação será excluída de forma permanente.</p>
-                                                            <p id="informacaoID"></p>
-                                                        </div>
-                                                        <div class="modal-footer">
-                                                            <form action="{{ route('user.destroy', $item->id) }}" method="POST">
-                                                                @csrf
-                                                                @method('delete')
-                                                                <button type="button" class="btn btn-light" data-bs-dismiss="modal">Cancelar</button>
-                                                                <button type="submit" class="btn btn-outline-light">Confirmar exclusão</button>
-                                                            </form>
-                                                        </div>
-                                                    </div>
-                                                </div>
+
+                                        @if ($item->id != Auth::user()->id)
+                                            <div style="display: inline-block" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="danger-tooltip" data-bs-title="Excluir registro">
+                                                <a href="javascript: deleteModal({{ $item->id }});" class="text-reset fs-16 px-1"> 
+                                                    <i class="ri-delete-bin-2-line"></i>
+                                                </a>
                                             </div>
-                                        </div> --}}
+                                        @endif                                        
+                                        
                                     </td>
                                 </tr>
                             @endforeach
@@ -93,7 +64,6 @@
         $('#modal-exclusao').modal('show');
     }
 </script>
-{{-- <button  type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#fill-danger-modal">Danger Filled</button> --}}
 
 <div id="modal-exclusao" class="modal fade" data-bs-backdrop="static" tabindex="-1" role="dialog" aria-labelledby="fill-danger-modalLabel" aria-hidden="true">
     <div class="modal-dialog">
